Fixed lingering problem in tests after merges of conflicting branches
